[0.4.4.0α] FEAT :: BETTER GRASS
Improved grass instancing through scalable zoning & tweaked looks with new sprite.

Body:
(+) Created grass_spawner and grass_spawn_handler scripts.
(*) Rewrote grass shader for better result.

Footer:
grass_spawner.gd instantiates each MeshInstance3D through the MultiMesh system.
It bases the area on which to instantiate the grass and number of meshes on the parameters passed to it by the handler.
grass_spawn_handler.gd calculates how many instances each area will instantiate, and where the center and size of each area is.
The idea is to get a granular customization of the amount of instances and groups for better balance between looks and performance.

from: [0.4.3.2α]